\name{Data_potato}
\alias{Data_potato}
\docType{data}
\title{
Monthly Average Potato Price of Delhi Market (India)
}
\description{
Monthly Average Potato Price of  Delhi Market from January 2010 to July 2020.
}
\usage{data("Data_potato")}
\format{
  A time series data with 127 observations.
  \describe{
    \item{\code{price}}{a time series}
  }
}
\details{
Dataset contains 127 observations of  monthly average potato price of  Delhi Market. It is obtained from Agmarknet Website.
}
\source{
https://agmarknet.gov.in/
}
\references{
https://agmarknet.gov.in/
}
\examples{
data(Data_potato)
}
\keyword{datasets}
